+++ src/libstore/unix/include/nix/store/user-lock.hh
@@ -7,6 +7,8 @@
 
 namespace nix {
 
+struct LocalSettings;
+
 struct UserLock
 {
     virtual ~UserLock() {}
@@ -36,8 +38,8 @@ struct UserLock
  * Acquire a user lock for a UID range of size `nrIds`. Note that this
  * may return nullptr if no user is available.
  */
-std::unique_ptr<UserLock> acquireUserLock(const std::string & userGroup, uid_t nrIds, bool useUserNamespace);
+std::unique_ptr<UserLock> acquireUserLock(const LocalSettings & localSettings, uid_t nrIds, bool useUserNamespace);
 
-bool useBuildUsers();
+bool useBuildUsers(const LocalSettings &);
 
 } // namespace nix
